Add command line argument for output directory
This commit adds a command line argument for specifying the output directory in the `SuperNetApp` class.
