+++ super_net/models/grid_pdf/grid_pdf/app.py
@@ -22,28 +22,6 @@ grid_pdf_providers = [
 class GridPdfApp(SuperNetApp):
     config_class = GridPdfConfig
 
-    @property
-    def argparser(self):
-        """Parser arguments for grid_pdf app can be added here"""
-        parser = super().argparser
-
-        parser.add_argument(
-            "-o",
-            "--output",
-            nargs="?",
-            default=None,
-            help="Name of the output directory.",
-        )
-
-        return parser
-
-    def get_commandline_arguments(self, cmdline=None):
-        """Get commandline arguments"""
-        args = super().get_commandline_arguments(cmdline)
-        if args["output"] is None:
-            args["output"] = pathlib.Path(args["config_yml"]).stem
-        return args
-
 
 def main():
     a = GridPdfApp(name="grid_pdf", providers=grid_pdf_providers)
