+++ super_net/models/wmin/wmin/app.py
@@ -22,28 +22,6 @@ wmin_providers = [
 class WminApp(SuperNetApp):
     config_class = WminConfig
 
-    @property
-    def argparser(self):
-        """Parser arguments for wmin app can be added here"""
-        parser = super().argparser
-
-        parser.add_argument(
-            "-o",
-            "--output",
-            nargs="?",
-            default=None,
-            help="Name of the output directory.",
-        )
-
-        return parser
-
-    def get_commandline_arguments(self, cmdline=None):
-        """Get commandline arguments"""
-        args = super().get_commandline_arguments(cmdline)
-        if args["output"] is None:
-            args["output"] = pathlib.Path(args["config_yml"]).stem
-        return args
-
 
 def main():
     a = WminApp(name="wmin", providers=wmin_providers)
